.archive/cleanup_20260318_072800/pyproject.toml
[project]
name = "project-spect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.53.0",
    "llama-cpp-python>=0.3.16",
    "mypy>=1.19.1",
    "openai>=2.8.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.10",
    "streamlit>=1.52.0",
    "types-pyyaml>=6.0.12.20250915",
    "uvicorn>=0.40.0",
    "litellm>=1.82.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["core"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
